% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUCAgg.R
\name{combine_multis}
\alias{combine_multis}
\title{Combine HUC polygons}
\usage{
combine_multis(subhucPoly)
}
\arguments{
\item{subhucPoly}{A subset of HUCs with an AREASQKM and HUC12 field.}
}
\value{
A deduplicated subhucPoly SpatialPolygonsDataFrame
}
\description{
A convenience function to combine HUCs stored as single polygons into multipolygons
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
