% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromHUC_finder.R
\name{fromHUC_finder}
\alias{fromHUC_finder}
\title{FromHUC Finder}
\usage{
fromHUC_finder(huc, hucs, tohucs)
}
\arguments{
\item{huc}{The huc in question}

\item{hucs}{The list of all HUCs}

\item{tohucs}{The list of TUHUCs}
}
\value{
The list of fromHUCs
}
\description{
This function finds the hucs that flow into a given huc.
}
\examples{
load(system.file("extdata","testhucpoly.rda",package="HUCAgg"))
print(testhucPoly@data$HUC12[which(testhucPoly@data$TOHUC \%in\% "070900020504")])
print(testhucPoly@data$HUC12[which(testhucPoly@data$TOHUC \%in\% "070900020604")])
hucList<-testhucPoly@data$HUC12
fromHUC<-sapply(hucList,fromHUC_finder,hucs=testhucPoly@data$HUC12,tohucs=testhucPoly@data$TOHUC)
print(unlist(fromHUC["070900020504"][[1]]))
print(unlist(fromHUC["070900020604"][[1]]))

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

