% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHUCList.R
\name{getHUCList}
\alias{getHUCList}
\title{Get list of HUCs}
\usage{
getHUCList(subRegion, subhucPoly)
}
\arguments{
\item{subRegion}{A HUC 02 or 04 for which to retreive a list of all HUC12s.}

\item{subhucPoly}{A subset of HUCs for the region in question.}
}
\value{
A list of hucs found.
}
\description{
A convenience function to retreive lists of HUCs according to a regex.
}
\examples{
\dontrun{
load(system.file("extdata","testhucpoly.rda",package="HUCAgg"))
print(testhucPoly@data$HUC12)
hucList<-as.character(unlist(getHUCList("0709000205",testhucPoly)))
print(hucList)
}

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

