% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUC_aggregator.R
\name{HUC_aggregator}
\alias{HUC_aggregator}
\title{HUC Aggregator}
\usage{
HUC_aggregator(huc, fromHUC)
}
\arguments{
\item{huc}{The huc in question}

\item{fromHUC}{The list of fromHUCs for all the hucs}
}
\value{
The list of all fromHUCs upstream of given huc
}
\description{
Aggregates fromHUCs recursively.
}
\examples{
TBD

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

