% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUC_TS_aggregator.R
\name{HUC_TS_aggregator}
\alias{HUC_TS_aggregator}
\title{Aggregate HUC time series data}
\usage{
HUC_TS_aggregator(upstream_size, fromHUC, huc12_areaDF, huc12agg_areaDF, dataF)
}
\arguments{
\item{upstream_size}{Named integer list of the number of upstream HUCs for each huc.}

\item{fromHUC}{Named list of lists containing the HUCs that flow into each HUC.}

\item{huc12_areaDF}{\code{data.frame} with a single observation of area for each local HUC.}

\item{huc12agg_areaDF}{\code{data.frame} with a single observation of area for each aggregated HUC.}

\item{dataF}{\code{data.frame} with a set of observations for each local HUC.}
}
\value{
dataF \code{data.frame} with the set of observations aggregated upstream using a weighted average.
}
\description{
Aggregate time series data for a collection of HUCs
}
\examples{
\dontrun{
load(system.file("extdata","HUC_TS_aggregator_test.rda",package="HUCAgg"))
outData<-HUC_TS_aggregator(upstream_size,fromHUC,huc12_areaDF,huc12agg_areaDF,dataF)
}
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
