% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUCAgg.R
\docType{package}
\name{HUCAg-package}
\alias{HUCAg-package}
\title{HUCAgg}
\description{
\tabular{ll}{
Package: \tab HUCAgg\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\details{
This package is used to aggregate HUC12 geospatial and time series 
data to total upstream watersheds.
}

