% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_regions.R
\name{init_regions}
\alias{init_regions}
\title{Initialize HUC Region Files}
\usage{
init_regions(WBDPath, regionsPath)
}
\arguments{
\item{WBDPath}{The path to the WBD shapefile}

\item{regionsPath}{The path where the regional subset rda files should go}
}
\value{
A list of regions and sub-processing units.
}
\description{
Builds regional WBD files from a national shapefile. The shapefile can be found here:
ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/Staged/WBD/
}
\examples{
\dontrun{
WBDPath<-"/data/WBDHU12.shp"
regionsPath<-"regions"
regions<-init_regions(WBDPath, regionsPath)
}
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

