% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parHUC_aggTime.R
\name{parHUC_aggTime}
\alias{parHUC_aggTime}
\title{Parallel Weighted Average for HUCs}
\usage{
parHUC_aggTime(range, aggrHUCs, areas, aggregate_hucdata)
}
\arguments{
\item{range}{A vector of length two specifying which hucs to process}

\item{aggrHUCs}{The output of running fromHUC_finder for the entire dataset}

\item{areas}{The areas for all the huc12s}

\item{aggregate_hucdata}{The function for use in a newly spawned environment}

\item{dataF}{A data frame containing time series for each local huc}
}
\value{
1 for success the range and error condition for failure.
}
\description{
This is a convenience function to be called by parApply
}
\examples{
TBD
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

