% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parUnionHUCRunner.R
\name{parUnionHUCRunner}
\alias{parUnionHUCRunner}
\title{Parallel Union HUCs}
\usage{
parUnionHUCRunner(range, aggrHUCs, hucPoly, unionHUC, outPath)
}
\arguments{
\item{range}{A vector of length two specifying which hucs to process}

\item{aggrHUCs}{The output of running fromHUC_finder for the entire dataset}

\item{hucPoly}{An imported shapefile with all the HUC geometry}

\item{unionHUC}{The unionHUC function for use in a newly spawned environment}

\item{outPath}{The path for the newly spawned environment to write to}
}
\value{
1 for success the range and error condition for failure.
}
\description{
This is a convenience function to be called by parApply
}
\examples{
TBD

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

