% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyHucs.R
\name{simplifyHucs}
\alias{simplifyHucs}
\title{Simplify HUCs}
\usage{
simplifyHucs(subhucPoly, coordThresh = 50000, simpTol = 5e-05)
}
\arguments{
\item{subhucPoly}{A subset of HUCs for the region in questioned.}

\item{coordThresh}{A threshold number of nodes above which polygons will be simplified.}

\item{simpTol}{A tolerance to pass into the Douglas-Peuker algorith.}
}
\value{
The subhucPoly data frame with large polygons simplified.
}
\description{
A convenience function to help simplify very large aggregate HUCs.
}
\examples{
TBD

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

