% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unionHUC.R
\name{unionHUC}
\alias{unionHUC}
\title{Union HUC}
\usage{
unionHUC(huc, upstreamHUCs, hucPoly)
}
\arguments{
\item{huc}{The huc in question}

\item{upstreamHUCs}{A list of HUCs to be aggregated into one geometry}

\item{hucPoly}{An imported shapefile with all the HUC geometry}
}
\value{
An aggregated polygon
}
\description{
Aggregates geometry for a set of HUCs
}
\examples{
TBD
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

