% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionHUCSet.R
\name{unionHUCSet}
\alias{unionHUCSet}
\title{Union HUC Set}
\usage{
unionHUCSet(aggrHUCs, fromHUCs, subhucPoly)
}
\arguments{
\item{aggrHUCs}{The list of aggregate HUCs to be processed per HUC_aggregator}

\item{fromHUCs}{The list of 'fromHUCs' to be processed per fromHUC_finder}

\item{subhucPoly}{A subset of HUCs for the region in questioned.}
}
\value{
The subhucPoly data frame with requested aggregated HUCs unioned together. 
Note that Polygon 'ID's are set to the HUC id and areas are added up appropriately.
}
\description{
Aggregates geometry for a large collection of HUCs. Note that subhucPoly must be all potential HUCs required.
}
\examples{
load(system.file("extdata","testhucpoly.rda",package="HUCAgg"))
plot(testhucPoly)
hucList<-as.character(unlist(getHUCList("07",testhucPoly)))
fromHUC<-sapply(hucList,fromHUC_finder,hucs=testhucPoly@data$HUC12,tohucs=testhucPoly@data$TOHUC)
aggrHUCs<-sapply(hucList, HUC_aggregator, fromHUC=fromHUC)
testhucPoly<-unionHUCSet(aggrHUCs, fromHUC, testhucPoly)
plot(subset(testhucPoly,testhucPoly@data$HUC12 \%in\% "070900020904"), add=TRUE, col=rgb(1,0,0,.3))
plot(subset(testhucPoly,testhucPoly@data$HUC12 \%in\% "070900020702"), add=TRUE, col=rgb(1,1,0,.3))

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

