% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromHUC_finder.R
\name{fromHUC_finder}
\alias{fromHUC_finder}
\title{FromHUC Finder}
\usage{
fromHUC_finder(huc, hucs, tohucs)
}
\arguments{
\item{huc}{The huc in question}

\item{hucs}{The list of all HUCs}

\item{tohucs}{The list of TUHUCs}
}
\value{
The list of fromHUCs
}
\description{
This function finds the hucs that flow into a given huc.
}
\examples{
TBD

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

