% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUC_aggregator.R
\name{HUC_aggregator}
\alias{HUC_aggregator}
\title{HUC Aggregator}
\usage{
HUC_aggregator(huc, fromHUC)
}
\arguments{
\item{huc}{The huc in question}

\item{fromHUC}{The list of fromHUCs for all the hucs}
}
\value{
The list of all fromHUCs upstream of given huc
}
\description{
Aggregates fromHUCs recursively.
}
\examples{
load(system.file("extdata","testhucpoly.rda",package="HUCAgg"))
hucList<-testhucPoly@data$HUC12
fromHUC<-sapply(hucList,fromHUC_finder,
                hucs=testhucPoly@data$HUC12,
                tohucs=testhucPoly@data$TOHUC)
print(unlist(fromHUC["070900020604"][[1]]))
print(c(unlist(fromHUC["070900020504"][[1]]), 
        unlist(fromHUC["070900020602"][[1]]), 
        unlist(fromHUC["070900020603"][[1]])))
aggrHUCs<-sapply(hucList, HUC_aggregator, fromHUC=fromHUC)
print(unlist(aggrHUCs["070900020604"][[1]]))

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

