% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionHUC.R
\name{unionHUC}
\alias{unionHUC}
\title{Union HUC}
\usage{
unionHUC(huc, upstreamHUCs, hucPoly)
}
\arguments{
\item{huc}{The huc in question}

\item{upstreamHUCs}{A list of HUCs to be aggregated into one geometry}

\item{hucPoly}{An imported shapefile with all the HUC geometry}
}
\value{
An aggregated polygon
}
\description{
Aggregates geometry for a set of HUCs.
}
\examples{
\dontrun{
load(system.file("extdata","testhucpoly.rda",package="HUCAgg"))
plot(testhucPoly)
hucList<-testhucPoly@data$HUC12
fromHUC<-sapply(hucList,fromHUC_finder,hucs=testhucPoly@data$HUC12,tohucs=testhucPoly@data$TOHUC)
aggrHUCs<-sapply(hucList, HUC_aggregator, fromHUC=fromHUC)
huc<-"070900020904"
outhucPoly<-unionHUC(huc, aggrHUCs, testhucPoly)
plot(outhucPoly, add=TRUE, col=rgb(1,0,0,.3))
}

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
